backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.28.0
numpy==1.26.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
pytest==8.0.2
httpx==0.27.0
python-multipart==0.0.9
